only users who created posts can edit and delete posts

+++ server/routes/posts.js
@@ -36,7 +36,8 @@ router.post(
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + '/images/' + req.file.filename
+        imagePath: url + '/images/' + req.file.filename,
+        creator: req.userData.userId
     });
     post.save().then(createdPost => {
         res.status(201).json({
@@ -103,12 +104,19 @@ router.put(
         _id: req.body.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath: imagePath
+        imagePath: imagePath,
+        creator: req.userData.userId
     });
-    Post.updateOne({ _id: req.params.id }, post).then(() => {
-        res.status(200).json({
-            message: 'Post updated.'
-        });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then((result) => {
+        if (result.nModified > 0) {
+            res.status(200).json({
+                message: 'Post updated successfully.'
+            });
+        } else {
+            res.status(401).json({
+                message: 'Post update failed.'
+            });
+        }
     });
 });
 
@@ -116,10 +124,16 @@ router.delete(
     '/:id', 
     checkAuth,
     (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id }).then(() => {
-        res.status(200).json({
-            message: 'Post deleted.'
-        });
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then((result) => {
+        if (result.n > 0) {
+            res.status(200).json({
+                message: 'Post deleted successfully.'
+            });
+        } else {
+            res.status(401).json({
+                message: 'Post deletion failed.'
+            });
+        }
     });
 });
